Add a new test into bonds test
I add a new test which is used to calculate how much fraction should be invest into a portfolio in order to gain a particular return despiete of the change of yield to maturity.

The user can input the basic imformation of two portfolio and the duration that he want. And the test that I add can automatically calculate  how much should be invest into each bonds.